Fix an issue with Save-for-Later alert not shown when saving from the feed
